changed: move ProgramDatabase inside programs folder
