+++ xbmc/programs/ProgramDatabase.cpp
