+++ xbmc/programs/ProgramDatabase.h
